# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AlphaAI Backend API",
    "version": "1.0.0",
    "description": "API documentation for the AlphaAI Backend service",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key needed to access the endpoints"
      }
    },
    "schemas": {
...
      "ContextSnippet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "companyId": {
            "type": "integer"
          },
          "personId": {
            "type": "integer"
          },
          "companyValueProp": {
            "type": "string"
          },
          "productNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pricingModel": {
            "type": "string"
          },
          "keyCompetitors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "companyDomain": {
            "type": "string"
          },
          "emailDomain": {
            "type": "string"
          },
          "topLinks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "companyId",
          "personId",
          "companyValueProp",
          "productNames",
          "pricingModel",
          "keyCompetitors",
          "companyDomain",
          "topLinks"
        ]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/snippets/company/{company_id}": {
      "get": {
        "summary": "Get snippets by company ID",
        "tags": [
          "Snippets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "responses": {
          "200": {
            "description": "List of context snippets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContextSnippet"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No snippets found for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
...
}